Move some datatypes from `.Offchain` up to `.Monads`
